Updated scripts

+++ make_sd.sh
@@ -18,12 +18,10 @@ if [[ "Darwin" == "$(uname -s)" ]]; then
     ln -s "/Volumes/NO NAME" ./mnt
     rm -rf ./mnt/*
 else
-    DEVICE_SIZE=$(blockdev --getsz "/dev/${DEVICE}")
-
+    #DEVICE_SIZE=$(blockdev --getsz "/dev/${DEVICE}")
     PARTITIONS="$(fdisk -l "/dev/${DEVICE}" | grep "^/dev/${DEVICE}")"
 
     echo "${PARTITIONS}"
-
 
     mkdir -p "${DIR_THIS}/mnt"
 
@@ -36,7 +34,7 @@ else
     mount "/dev/${BOOT_DEVICE}" "${DIR_THIS}/mnt"
 fi
 
-pushd ${DIR_THIS}/mnt
+pushd "${DIR_THIS}/mnt"
 tar xf "${DIR_THIS}/alpine-rpi-${ALPINE_VERSION}-${ALPINE_ARCH}.tar.gz" --no-same-owner  
 popd
 
@@ -45,19 +43,17 @@ DIR_APKOVL="${DIR_THIS}/apkovl"
 if [[ ! -d "${DIR_APKOVL}" ]]; then
     mkdir -p "${DIR_APKOVL}"
 else
-    rm -rf "${DIR_APKOVL}/"*
+    rm -rf "${DIR_APKOVL:?}/"*
 fi
 
 cp -a "${DIR_THIS}/src/"* "${DIR_APKOVL}"
 
-APKREPOSOPTS=-f
-#APKREPOSOPTS=-1
-#APKREPOSOPTS=mirror1.hs-esslingen.de
-
-echo "BOOT_DEVICE_NAME=${BOOT_DEVICE}" >> "${DIR_APKOVL}/mypi-setup/config"
-echo "ROOT_DEVICE_NAME=${ROOT_DEVICE}" >> "${DIR_APKOVL}/mypi-setup/config"
-echo "WLAN_SSID=${WLAN_SSID}"          >> "${DIR_APKOVL}/mypi-setup/config"
-echo "WLAN_PASSWORD=${WLAN_PASSWORD}"  >> "${DIR_APKOVL}/mypi-setup/config"
+cat > "${DIR_APKOVL}/mypi-setup/config" <<-__EOF__
+BOOT_DEVICE_NAME=${BOOT_DEVICE}"
+ROOT_DEVICE_NAME=${ROOT_DEVICE}"
+WLAN_SSID=${WLAN_SSID}"
+WLAN_PASSWORD=${WLAN_PASSWORD}"
+__EOF__
 
 cat > "${DIR_APKOVL}/mypi-setup/answer.txt" <<-__EOF__
 KEYMAPOPTS="us us"
@@ -71,87 +67,12 @@ iface eth0 inet dhcp
 "
 TIMEZONEOPTS="-z UTC"
 PROXYOPTS=none
-APKREPOSOPTS="${APKREPOSOPTS}"
+APKREPOSOPTS="-f"
 SSHDOPTS="-c openssh"
 NTPOPTS="-c busybox"
 APKCACHEOPTS="none"
 LBUOPTS="none"
 __EOF__
-
-###############################################################################
-#
-#                                                      /<BOOT>/setup-phase-2.sh
-#
-(
-cat <<-__EOF__
-#!/bin/sh
-
-set -x
-set -e
-
-rc-update del hwclock boot &> /dev/null|| true
-rc-update add swclock boot
-
-apk add bash htop jq docker perl
-
-rc-update add docker boot
-service docker start
-__EOF__
-
-if [[ -n "${WLAN_SSID}" && -n "${WLAN_PASSWORD}" ]]; then
-cat <<-__EOF__
-apk add wireless-tools wpa_supplicant
-ip link | grep wlan0
-ip link set wlan0 up
-iwlist wlan0 scanning
-iwconfig wlan0 essid ${WLAN_SSID}
-
-cat > /etc/wpa_supplicant/wpa_supplicant.conf <<-_EOF_
-ctrl_interface=DIR=/var/run/wpa_supplicant GROUP=netdev
-update_config=1
-_EOF_
-
-wpa_passphrase "${WLAN_SSID}" "${WLAN_PASSWORD}" > /etc/wpa_supplicant/wpa_supplicant.conf
-wpa_supplicant -B -i wlan0 -c /etc/wpa_supplicant/wpa_supplicant.conf
-udhcpc -i wlan0
-ip addr show wlan0
-
-###############################################################################
-#
-# Update /etc/network/interfaces
-#
-cat >> /etc/network/interfaces <<-_EOF_
-
-auto wlan0
-iface wlan0 inet dhcp
-wpa-conf /etc/wpa_supplicant/wpa_supplicant.conf
-_EOF_
-
-###############################################################################
-#
-# Configure udhcpc to ignore DNS and GATEWAY from wlan0
-#
-mkdir -p /etc/udhcpc
-cat > /etc/udhcpc/udhcpc.conf <<-_EOF_
-NO_DNS=wlan0
-NO_GATEWAY=wlan0
-_EOF_
-
-###############################################################################
-#
-# Enable WLAN
-#
-ifconfig wlan0 down
-/etc/init.d/wpa_supplicant start
-rc-update add wpa_supplicant boot
-
-__EOF__
-fi
-)  > "${DIR_THIS}/mnt/setup-phase-2.sh"
-#
-#                                                      /<BOOT>/setup-phase-2.sh
-#
-###############################################################################
 
 pushd "${DIR_APKOVL}"
     tar czf ${DIR_THIS}/mnt/${RPI_HOSTNAME}.apkovl.tar.gz .
